# docs/swagger.yaml
definitions:
  app.CollectDefectsDTO:
    properties:
      component:
        type: string
      issue_id:
        type: string
      score:
        type: string
      status:
        type: string
      version:
        type: string
    type: object
  controller.ResponseData:
    properties:
      code:
        type: string
      data: {}
      msg:
        type: string
    type: object
  controller.defectRequest:
    properties:
      abi:
        type: string
      affected_version:
        items:
          type: string
        type: array
      component:
        type: string
      description:
        type: string
      guidance_url:
        type: string
      influence:
        type: string
      issue_number:
        type: string
      issue_org:
        type: string
      issue_repo:
        type: string
      issue_status:
        type: string
      kernel:
        type: string
      reference_url:
        type: string
      severity_level:
        type: string
      system_version:
        type: string
    required:
    - abi
    - affected_version
    - component
    - description
    - guidance_url
    - influence
    - issue_number
    - issue_org
    - issue_repo
    - issue_status
    - kernel
    - reference_url
    - severity_level
    - system_version
    type: object
info:
  contact: {}
paths:
  /v1/defect:
    get:
      consumes:
      - application/json
      description: collect information of some defects
      parameters:
      - description: collect defects after the date
        in: query
        name: date
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/app.CollectDefectsDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: collect information of some defects
      tags:
      - Defect
    post:
      consumes:
      - application/json
      description: add or update a defect
      parameters:
      - description: body of a defect
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/controller.defectRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: add or update a defect
      tags:
      - Defect
  /v1/defect/bulletin:
    post:
      consumes:
      - application/json
      description: generate security bulletin for some defects
      parameters:
      - description: body of some issue number
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/controller.defectRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: generate security bulletin for some defects
      tags:
      - Defect
swagger: "2.0"